+++ src/components/__tests__/default-button-props.js
@@ -1,0 +1,9 @@
+const defaultValues = {
+  Theme: { getCurrent: () => ({ properties: { palettes: { ui: [{ color: ['color1', 'color2'] }] } } }) },
+  layout: { style: { label: 'testLabel' } },
+  button: { setAttribute: () => {} },
+  engineApp: {},
+  context: { permissions: ['interact'] },
+};
+
+export default defaultValues;
